Combining services to one folder, contact-us uses env varibles
