+++ .env.example
@@ -1,2 +1,3 @@
 # Additional env vars are constructed in bin/load-env.sh
 ENVIRONMENT_NAME="YOUR-NAME-GOES-HERE"
+CONTACT_US_EMAIL_TO="YOUR-EMAIL-GOES-HERE"
