+++ bin/ci-deploy.sh
@@ -15,7 +15,7 @@ createCommitSite() {
   make build
   echo Copying assets to S3
   aws s3 sync ./dist/assets s3://$BUCKET_NAME/$COMMIT_REF/assets
-  make publish-service-deploy
+  make services-deploy
   make publish-service-invoke
   echo Done!
