--- contact-us-service/config/env.example.json
@@ -1,6 +1,0 @@
-{
-  "emailTo": [
-    "hello@red-badger.com"
-  ],
-  "emailFrom": "hello@red-badger.com"
-}
