--- contact-us-service/event.json
@@ -1,4 +1,0 @@
-{
-  "message": "Hi I want to work with you",
-  "contact": "zoe.bryant@red-badger.com"
-}
