--- contact-us-service/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "contact-us-service",
-  "version": "1.0.0",
-  "description": "Contact us service",
-  "main": "handler.js",
-  "scripts": {
-    "test": "sls deploy && sls invoke --function contactUs -p event.json"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "aws-sdk": "2.4.9"
-  }
-}
