--- contact-us-service/serverless.yml
@@ -1,23 +1,0 @@
-service: contact-us-service
-
-provider:
-  name: aws
-  runtime: nodejs4.3
-  region: eu-west-1
-  iamRoleStatements:
-    - Effect: 'Allow'
-      Action:
-        - 'SES:SendEmail'
-        - 'SES:SendRawEmail'
-      Resource: 'arn:aws:ses:eu-west-1:578418881509:identity/hello@red-badger.com'
-
-functions:
-  contactUs:
-    handler: handler.contactUs
-    events:
-      - http:
-          path: contact-us
-          method: post
-          request:
-            template:
-              application/json: "{\"message\" : \"$input.path('$.message')\",\"contact\" : \"$input.path('$.contact')\"}"
